+++ finstack/valuations/tests/instruments/bond/integration/mod.rs
@@ -4,6 +4,4 @@
 mod amortizing_bonds;
 mod callable_bonds;
 mod custom_structures;
-mod quantlib_parity;
-mod quantlib_parity_callable;
 mod tree_calibration_validation;